Harris-Detect Demo
This demonstrates all of the most crucial steps necessary for the Harris Detection Algorithm. This is not meant to be a final implementation of the algorithm (firstly, because separating it into separate functions like this slows it down dramatically, and secondly, because the normalizations used in each function make them ideal for visualization but lead to unrealistic response values).

Later, we will roll up all of the functions into a single function called response().