Full Harris corner detection algorithm
This is the complete Harris corner detection algorithm. The response function rates pixels based on how "point-like" they are. That is, it looks at each pixel and its neighbors, and looks at how the values of the pixels change within the neighborhood of the pixel as one moves away from the main pixel. Pixels with large changes are more point-like, those with medium changes are edge-like, and those with little change are background-like. The program outputs a visualization of this function by displaying point-like pixels as white, and lower-valued pixels as black. The program can also be modified to display a red dot at the highest-valued pixel.